backend/pyproject.toml
[tool.poetry]
name = "helioscta_gas_ebbs_backend"
version = "0.1.0"
description = "Backend API for Helios CTA Gas EBBs"
authors = ["Your Name <your.email@example.com>"]
packages = [{include = "src"}]
exclude = ["*.logs", "*.png", "*.csv", "*.xlsx", "*.xlsm"]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
fastapi = ">=0.115.0"
uvicorn = {version = ">=0.34.0", extras = ["standard"]}
psycopg2-binary = "==2.9.10"
python-dotenv = "==1.0.1"
pandas = ">=2.0.0"
pyarrow = ">=12.0.0"
pytz = "*"
tabulate = "*"
requests = "*"
beautifulsoup4 = "*"

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
